+++ glean-core/android/src/main/java/mozilla/telemetry/glean/Dispatchers.kt
@@ -43,7 +43,10 @@ internal object Dispatchers {
                     }
                     null
                 }
-                else -> coroutineScope.launch(block = block)
+
+                else -> {
+                    coroutineScope.launch(block = block)
+                }
             }
     }