+++ glean-core/rlb/tests/health_ping_file_overwrite.rs
@@ -47,7 +47,7 @@ impl net::PingUploader for ReportingUploader {
 
 /// Test scenario: Write a client ID to the backup file and check that it's used after initialization.
 #[test]
-fn test_pre_post_init_health_pings_exist() {
+fn test_pre_init_health_ping_exist() {
     common::enable_test_logging();
 
     // Create a custom configuration to use a validating uploader.
@@ -79,7 +79,6 @@ fn test_pre_post_init_health_pings_exist() {
     let exception_uuid = &payload["metrics"]["uuid"]["glean.health_recovered_client_id"];
     assert_eq!(&JsonValue::Null, exception_uuid);
 
-    // TODO(bug 1996862): We don't run the mitigation yet.
-    //let ping_client_id = &payload["client_info"]["client_id"];
-    //assert_eq!(client_id, ping_client_id);
+    let ping_client_id = &payload["client_info"]["client_id"];
+    assert_eq!(client_id, ping_client_id);
 }