+++ glean-core/src/core/mod.rs
@@ -39,6 +39,11 @@ use crate::{
 const CLIENT_ID_PLAIN_FILENAME: &str = "client_id.txt";
 static GLEAN: OnceCell<Mutex<Glean>> = OnceCell::new();
 
+/// Rate limiting defaults
+/// 15 pings every 60 seconds.
+pub const DEFAULT_SECONDS_PER_INTERVAL: u64 = 60;
+pub const DEFAULT_PINGS_PER_INTERVAL: u32 = 15;
+
 pub fn global_glean() -> Option<&'static Mutex<Glean>> {
     GLEAN.get()
 }
@@ -202,8 +207,8 @@ impl Glean {
         // Create an upload manager with rate limiting of 15 pings every 60 seconds.
         let mut upload_manager = PingUploadManager::new(&cfg.data_path, &cfg.language_binding_name);
         let rate_limit = cfg.rate_limit.as_ref().unwrap_or(&PingRateLimit {
-            seconds_per_interval: 60,
-            pings_per_interval: 15,
+            seconds_per_interval: DEFAULT_SECONDS_PER_INTERVAL,
+            pings_per_interval: DEFAULT_PINGS_PER_INTERVAL,
         });
         upload_manager.set_rate_limiter(
             rate_limit.seconds_per_interval,
@@ -294,6 +299,11 @@ impl Glean {
             Ok(id) => Some(id),
             Err(ClientIdFileError::NotFound) => {
                 // That's ok, the file might just not exist yet.
+                glean
+                    .health_metrics
+                    .file_read_error
+                    .get("file-not-found")
+                    .add_sync(&glean, 1);
                 None
             }
             Err(ClientIdFileError::PermissionDenied) => {
@@ -331,14 +341,13 @@ impl Glean {
 
         {
             let data_store = glean.data_store.as_ref().unwrap();
-            let db_load_sizes = data_store.load_sizes.as_ref().unwrap();
-            let new_size = db_load_sizes.new.unwrap_or(0);
+            let file_size = data_store.file_size.map(|n| n.get()).unwrap_or(0);
 
             // If we have a client ID on disk, we check the database
             if let Some(stored_client_id) = stored_client_id {
                 // state (c)
-                if new_size <= 0 {
-                    log::trace!("no database. database size={new_size}. stored_client_id={stored_client_id}");
+                if file_size == 0 {
+                    log::trace!("no database. database size={file_size}. stored_client_id={stored_client_id}");
                     // state (d)
                     glean
                         .health_metrics
@@ -350,6 +359,10 @@ impl Glean {
                         .set_sync(&glean, ExceptionState::EmptyDb);
 
                     // state (e) -- mitigation: store recovered client ID in DB
+                    glean
+                        .core_metrics
+                        .client_id
+                        .set_from_uuid_sync(&glean, stored_client_id);
                 } else {
                     let db_client_id = glean
                         .core_metrics
@@ -366,6 +379,10 @@ impl Glean {
                                 .set_sync(&glean, ExceptionState::RegenDb);
 
                             // state (e) -- mitigation: store recovered client ID in DB
+                            glean
+                                .core_metrics
+                                .client_id
+                                .set_from_uuid_sync(&glean, stored_client_id);
                         }
                         Some(db_client_id) if db_client_id == *KNOWN_CLIENT_ID => {
                             // state (i)
@@ -383,6 +400,10 @@ impl Glean {
 
                             // If we have a recovered client ID we also overwrite the database.
                             // state (e)
+                            glean
+                                .core_metrics
+                                .client_id
+                                .set_from_uuid_sync(&glean, stored_client_id);
                         }
                         Some(db_client_id) if db_client_id == stored_client_id => {
                             // all valid. nothing to do
@@ -667,17 +688,6 @@ impl Glean {
             self.database_metrics
                 .rkv_load_error
                 .set_sync(self, rkv_load_state)
-        }
-
-        if let Some(load_sizes) = self
-            .data_store
-            .as_mut()
-            .and_then(|database| database.load_sizes())
-        {
-            self.database_metrics.load_sizes.set_sync(
-                self,
-                serde_json::to_value(load_sizes).unwrap_or(serde_json::json!({})),
-            );
         }
     }