+++ glean-core/src/internal_pings.rs
@@ -82,7 +82,7 @@ impl InternalPings {
                 true,
                 enabled,
                 vec![],
-                vec!["pre_init".to_string(), "post_init".to_string()],
+                vec!["pre_init".to_string()],
                 true,
                 vec![],
             ),