+++ glean-core/src/lib.rs
@@ -38,6 +38,10 @@ mod core_metrics;
 mod coverage;
 mod database;
 mod debug;
+#[cfg(feature = "benchmark")]
+#[doc(hidden)]
+pub mod dispatcher;
+#[cfg(not(feature = "benchmark"))]
 mod dispatcher;
 mod error;
 mod error_recording;
@@ -583,10 +587,6 @@ fn initialize_inner(
             // Now that Glean is initialized, we can capture the directory info from the pre_init phase and send it in
             // a health ping with reason "pre_init".
             record_dir_info_and_submit_health_ping(dir_info, "pre_init");
-
-            // Now capture a post_init snapshot of the state of Glean's data directories after initialization to send
-            // in a health ping with reason "post_init".
-            record_dir_info_and_submit_health_ping(collect_directory_info(data_path), "post_init");
 
             let state = global_state().lock().unwrap();
             if let Err(e) = state.callbacks.trigger_upload() {