+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v66.1.2...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v66.2.0...main)
+
+# v66.2.0 (2025-12-09)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v66.1.2...v66.2.0)
+
+* General
+  * Stop reporting db file sizes during init phase ([#3331](https://github.com/mozilla/glean/pull/3331))
+  * Tiny performance improvement for putting tasks on the dispatcher ([#3318](https://github.com/mozilla/glean/pull/3318))
+  * Instrument the case when the `client_id.txt` file does not exist yet ([#3339](https://github.com/mozilla/glean/pull/3339))
+  * When a missing client ID in the database is detected, Glean now restores the backup client ID ([#3334](https://github.com/mozilla/glean/pull/3334))
+  * Disable the `post_init` health ping ([#3343](https://github.com/mozilla/glean/pull/3343))
+* iOS
+  * Glean for iOS is now being built with Xcode 26.2 ([#3350](https://github.com/mozilla/glean/pull/3350))
+  * Fix strict concurrency warnings uncovered in Swift 6 ([#3325](https://github.com/mozilla/glean/pull/3325), [#3347](https://github.com/mozilla/glean/pull/3347))
+  * Correctly end background tasks ([#3347](https://github.com/mozilla/glean/pull/3347))
+* Android
+  *  Configuration-time performance optimizations for the Glean Gradle Plugin ([#3326](https://github.com/mozilla/glean/pull/3326))
+* Python
+  * Bindings are generated in library mode, enabling source install from PyPI ([#3317](https://github.com/mozilla/glean/pull/3317))
 
 # v66.1.2 (2025-11-25)